Circuit changes from review comments.
