+++ gcs/constraints/circuit.hh
@@ -1,0 +1,10 @@
+#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CIRCUIT_HH
+#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CIRCUIT_HH
+
+#include <gcs/constraints/circuit/circuit_scc.hh>
+
+namespace gcs
+{
+    using Circuit = CircuitSCC;
+}
+#endif // GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CIRCUIT_HH
