+++ gcs/constraints/circuit/circuit_no_backedges_test.cc
@@ -3,8 +3,7 @@
 
 #include <iostream>
 
-#include <boost/program_options.hpp>
-#include <gcs/constraints/circuit/circuit.hh>
+#include <gcs/constraints/circuit.hh>
 #include <gcs/constraints/in.hh>
 
 using namespace gcs;
@@ -17,8 +16,6 @@ using std::make_optional;
 using std::nullopt;
 using std::string;
 using std::vector;
-
-namespace po = boost::program_options;
 
 auto post_constraints(Problem & p, vector<IntegerVariableID> & nodes)
 {
@@ -42,43 +39,8 @@ auto post_constraints(Problem & p, vector<IntegerVariableID> & nodes)
     }
 }
 
-auto main(int argc, char * argv[]) -> int
+auto main(int, char *[]) -> int
 {
-    po::options_description display_options{"Program options"};
-    display_options.add_options()            //
-        ("help", "Display help information") //
-        ("prove", "Create a proof");         //
-
-    po::options_description all_options{"All options"};
-
-    //    all_options.add_options()(
-    //        "n", po::value<int>()->default_value(3), "Integer value n.") //
-    //        ;
-    all_options.add(display_options);
-    po::variables_map options_vars;
-
-    try {
-        po::store(po::command_line_parser(argc, argv)
-                      .options(all_options)
-                      .run(),
-            options_vars);
-        po::notify(options_vars);
-    }
-    catch (const po::error & e) {
-        cerr << "Error: " << e.what() << endl;
-        cerr << "Try " << argv[0] << " --help" << endl;
-        return EXIT_FAILURE;
-    }
-
-    if (options_vars.contains("help")) {
-        cout << "Usage: " << argv[0] << " [options] [size]" << endl;
-        cout << endl;
-        cout << display_options << endl;
-        return EXIT_SUCCESS;
-    }
-
-    //    auto n = options_vars["n"].as<int>();
-
     Problem p;
     auto nodes = p.create_integer_variable_vector(8, 0_i, 7_i);
 
@@ -109,8 +71,6 @@ auto main(int argc, char * argv[]) -> int
             return true;
         }},
         make_optional<ProofOptions>("scc_no_backedges.opb", "scc_no_backedges.veripb"));
-
-    // system("veripb --trace --useColor scc_no_backedges.opb scc_no_backedges.veripb");
 
     cout << stats;
