+++ gcs/constraints/circuit/circuit_prevent.hh
@@ -1,0 +1,20 @@
+#ifndef GLASGOW_CONSTRAINT_SOLVER_CIRCUIT_PREVENT_HH
+#define GLASGOW_CONSTRAINT_SOLVER_CIRCUIT_PREVENT_HH
+
+#include <gcs/constraints/circuit/circuit_base.hh>
+
+namespace gcs
+{
+    /**
+     * Circuit constraint that propagates by identifying chains and removing the head of each chain from the domain of
+     * the tail "preventing" small cycles.
+     */
+    class CircuitPrevent : public CircuitBase
+    {
+    public:
+        using CircuitBase::CircuitBase;
+        [[nodiscard]] auto clone() const -> std::unique_ptr<Constraint> override;
+        auto install(innards::Propagators &, innards::State &) && -> void override;
+    };
+}
+#endif // GLASGOW_CONSTRAINT_SOLVER_CIRCUIT_PREVENT_HH
