+++ test/CMakeLists.txt
@@ -1,15 +1,7 @@
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
 add_subdirectory(break_table)
 add_subdirectory(break_table_with_constants)
-add_subdirectory(break_temporary)
 add_subdirectory(break_views)
 add_subdirectory(up_proof)
 add_subdirectory(consolidate_unary)
-add_subdirectory(cpaior_circuit_experiments)
-add_subdirectory(disconnected)
-add_subdirectory(scc_prune_root)
-add_subdirectory(scc_prune_skip)
-add_subdirectory(scc_multiple_sccs)
-add_subdirectory(scc_no_back_edges)
 add_subdirectory(unsat_optimise)
-add_subdirectory(verify_circuit_proofs)
