--- test/break_temporary/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_executable(break_temporary break_temporary.cc)
-target_link_libraries(break_temporary PUBLIC glasgow_constraint_solver)
-
-find_package(Boost 1.74 COMPONENTS program_options REQUIRED)
-include_directories(${Boost_INCLUDE_DIR})
-target_link_libraries(break_temporary LINK_PUBLIC ${Boost_LIBRARIES})
