--- test/cpaior_circuit_experiments/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_executable(cpaior_circuit_experiments cpaior_circuit_experiments.cc)
-target_link_libraries(cpaior_circuit_experiments PUBLIC glasgow_constraint_solver)
-
-find_package(Boost 1.74 COMPONENTS program_options REQUIRED)
-include_directories(${Boost_INCLUDE_DIR})
-target_link_libraries(cpaior_circuit_experiments LINK_PUBLIC ${Boost_LIBRARIES})
