--- test/verify_circuit_proofs/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_executable(verify_circuit_proofs verify_circuit_proofs.cc)
-target_link_libraries(verify_circuit_proofs PUBLIC glasgow_constraint_solver)
-
-find_package(Boost 1.74 COMPONENTS program_options REQUIRED)
-include_directories(${Boost_INCLUDE_DIR})
-target_link_libraries(verify_circuit_proofs LINK_PUBLIC ${Boost_LIBRARIES})
